Convert to a more user-friendly configuration API
Each subsystem now specifies its defaults in its constructor, which writes
those defaults to NetworkTables. If a default value specified in the code does
not match what is found in NetworkTables, a notification is printed so that
the code can be updated.

Each access to retrieve a configuration value no longer requires specifying a
default value, as (briefly) demonstrated in ExampleSubsystem.java.